In this, we changed the Student class as below:
1- No duplicate course is allowed
2- Change the double type of grade to Float
3- Change the int type of section in 'Student.CourseSection' to Integer.

+++ src/domain/EnrollCtrl.java
@@ -7,10 +7,10 @@ import domain.exceptions.EnrollmentRulesViolationException;
 
 public class EnrollCtrl {
 	public void enroll(Student s, List<CSE> courses) throws EnrollmentRulesViolationException {
-        Map<Term, Map<Course, Double>> transcript = s.getTranscript();
+        Map<Term, Map<Course, Float>> transcript = s.getTranscript();
 		for (CSE o : courses) {
-            for (Map.Entry<Term, Map<Course, Double>> tr : transcript.entrySet()) {
-                for (Map.Entry<Course, Double> r : tr.getValue().entrySet()) {
+            for (Map.Entry<Term, Map<Course, Float>> tr : transcript.entrySet()) {
+                for (Map.Entry<Course, Float> r : tr.getValue().entrySet()) {
                     if (r.getKey().equals(o.getCourse()) && r.getValue() >= 10)
                         throw new EnrollmentRulesViolationException(String.format("The student has already passed %s", o.getCourse().getName()));
                 }
@@ -18,8 +18,8 @@ public class EnrollCtrl {
 			List<Course> prereqs = o.getCourse().getPrerequisites();
 			nextPre:
 			for (Course pre : prereqs) {
-                for (Map.Entry<Term, Map<Course, Double>> tr : transcript.entrySet()) {
-                    for (Map.Entry<Course, Double> r : tr.getValue().entrySet()) {
+                for (Map.Entry<Term, Map<Course, Float>> tr : transcript.entrySet()) {
+                    for (Map.Entry<Course, Float> r : tr.getValue().entrySet()) {
                         if (r.getKey().equals(pre) && r.getValue() >= 10)
                             continue nextPre;
                     }
@@ -40,8 +40,8 @@ public class EnrollCtrl {
 			unitsRequested += o.getCourse().getUnits();
 		double points = 0;
 		int totalUnits = 0;
-        for (Map.Entry<Term, Map<Course, Double>> tr : transcript.entrySet()) {
-            for (Map.Entry<Course, Double> r : tr.getValue().entrySet()) {
+        for (Map.Entry<Term, Map<Course, Float>> tr : transcript.entrySet()) {
+            for (Map.Entry<Course, Float> r : tr.getValue().entrySet()) {
                 points += r.getValue() * r.getKey().getUnits();
                 totalUnits += r.getKey().getUnits();
             }
